Move `EA.RazorCompiler` to `Experimental`
